moving azure resources and permission code to azure module under resources and access and moving the related test files
