+++ src/databricks/labs/ucx/cli.py
@@ -12,7 +12,7 @@ from databricks.sdk.errors import NotFound
 
 from databricks.labs.ucx.account import AccountWorkspaces, WorkspaceInfo
 from databricks.labs.ucx.assessment.aws import AWSResourcePermissions
-from databricks.labs.ucx.assessment.azure import AzureResourcePermissions
+from databricks.labs.ucx.azure.access import AzureResourcePermissions
 from databricks.labs.ucx.config import WorkspaceConfig
 from databricks.labs.ucx.framework.crawlers import StatementExecutionBackend
 from databricks.labs.ucx.hive_metastore import ExternalLocations, TablesCrawler
