+++ tests/unit/assessment/test_azure.py
@@ -1,224 +1,9 @@
-import base64
-import json
-from unittest.mock import Mock, create_autospec
+from unittest.mock import Mock
 
-import pytest
-from databricks.labs.blueprint.installation import MockInstallation
-from databricks.sdk import WorkspaceClient
-from databricks.sdk.oauth import Token
-
-from databricks.labs.ucx.assessment.azure import (
-    AzureResource,
-    AzureResourcePermissions,
-    AzureResources,
-    AzureServicePrincipalCrawler,
-    Principal,
-)
-from databricks.labs.ucx.hive_metastore import ExternalLocations
+from databricks.labs.ucx.assessment.azure import AzureServicePrincipalCrawler
 
 from ..framework.mocks import MockBackend
-from . import get_az_api_mapping, workspace_client_mock
-
-
-@pytest.fixture
-def az_token(mocker):
-    token = json.dumps({"aud": "foo", "tid": "bar"}).encode("utf-8")
-    str_token = base64.b64encode(token).decode("utf-8").replace("=", "")
-    tok = Token(access_token=f"header.{str_token}.sig")
-    mocker.patch("databricks.sdk.oauth.Refreshable.token", return_value=tok)
-
-
-def test_subscriptions_no_subscription(mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    azure_resource = AzureResources(w, include_subscriptions="001")
-    subscriptions = list(azure_resource.subscriptions())
-    assert len(subscriptions) == 0
-
-
-def test_subscriptions_valid_subscription(mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    azure_resource = AzureResources(w, include_subscriptions="002")
-    subscriptions = list(azure_resource.subscriptions())
-    assert len(subscriptions) == 1
-    for subscription in subscriptions:
-        assert subscription.name == "sub2"
-
-
-def test_storage_accounts(mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    azure_resource = AzureResources(w, include_subscriptions="002")
-    storage_accounts = list(azure_resource.storage_accounts())
-    assert len(storage_accounts) == 2
-    for storage_account in storage_accounts:
-        assert storage_account.resource_group == "rg1"
-        assert storage_account.storage_account in ["sto2", "sto3"]
-        assert storage_account.container is None
-        assert storage_account.subscription_id == "002"
-
-
-def test_containers(mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    azure_resource = AzureResources(w, include_subscriptions="002")
-    azure_storage = AzureResource("subscriptions/002/resourceGroups/rg1/storageAccounts/sto2")
-    containers = list(azure_resource.containers(azure_storage))
-    assert len(containers) == 3
-    for container in containers:
-        assert container.resource_group == "rg1"
-        assert container.storage_account == "sto2"
-        assert container.container in ["container1", "container2", "container3"]
-        assert container.subscription_id == "002"
-
-
-def test_role_assignments_storage(mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    azure_resource = AzureResources(w, include_subscriptions="002")
-    resource_id = "subscriptions/002/resourceGroups/rg1/storageAccounts/sto2"
-    role_assignments = list(azure_resource.role_assignments(resource_id))
-    assert len(role_assignments) == 1
-    for role_assignment in role_assignments:
-        assert role_assignment.role_name == "Contributor"
-        assert role_assignment.principal == Principal("appIduser2", "disNameuser2", "Iduser2")
-        assert str(role_assignment.scope) == resource_id
-        assert role_assignment.resource == AzureResource(resource_id)
-
-
-def test_role_assignments_container(mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    azure_resource = AzureResources(w, include_subscriptions="002")
-    resource_id = "subscriptions/002/resourceGroups/rg1/storageAccounts/sto2/containers/container1"
-    role_assignments = list(azure_resource.role_assignments(resource_id))
-    assert len(role_assignments) == 1
-    for role_assignment in role_assignments:
-        assert role_assignment.role_name == "Contributor"
-        assert role_assignment.principal == Principal("appIduser2", "disNameuser2", "Iduser2")
-        assert str(role_assignment.scope) == resource_id
-        assert role_assignment.resource == AzureResource(resource_id)
-
-
-def test_save_spn_permissions_no_external_table(caplog):
-    w = create_autospec(WorkspaceClient)
-    rows = {"SELECT \\* FROM ucx.external_locations": []}
-    backend = MockBackend(rows=rows)
-    location = ExternalLocations(w, backend, "ucx")
-    installation = MockInstallation()
-    azure_resource_permission = AzureResourcePermissions(
-        installation, w, AzureResources(w, include_subscriptions="002"), location
-    )
-    azure_resource_permission.save_spn_permissions()
-    msg = "There are no external table present with azure storage account. Please check if assessment job is run"
-    assert [rec.message for rec in caplog.records if msg in rec.message]
-
-
-def test_save_spn_permissions_no_azure_storage_account():
-    w = create_autospec(WorkspaceClient)
-    rows = {"SELECT \\* FROM ucx.external_locations": [["s3://bucket1/folder1", "0"]]}
-    backend = MockBackend(rows=rows)
-    location = ExternalLocations(w, backend, "ucx")
-    installation = MockInstallation()
-    azure_resource_permission = AzureResourcePermissions(
-        installation, w, AzureResources(w, include_subscriptions="002"), location
-    )
-    storage_accounts = azure_resource_permission._get_storage_accounts()
-    assert len(storage_accounts) == 0
-
-
-def test_save_spn_permissions_valid_azure_storage_account():
-    w = create_autospec(WorkspaceClient)
-    rows = {
-        "SELECT \\* FROM ucx.external_locations": [
-            ["s3://bucket1/folder1", "1"],
-            ["abfss://continer1@storage1.dfs.core.windows.net/folder1", "1"],
-        ]
-    }
-    backend = MockBackend(rows=rows)
-    location = ExternalLocations(w, backend, "ucx")
-    installation = MockInstallation()
-    azure_resource_permission = AzureResourcePermissions(
-        installation, w, AzureResources(w, include_subscriptions="002"), location
-    )
-    storage_accounts = azure_resource_permission._get_storage_accounts()
-    assert storage_accounts[0] == "storage1"
-
-
-def test_map_storage_with_spn_no_blob_permission(mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    location = ExternalLocations(w, MockBackend(), "ucx")
-    resource_id = "subscriptions/002/resourceGroups/rg1/storageAccounts/sto3"
-    installation = MockInstallation()
-    azure_resource_permission = AzureResourcePermissions(
-        installation, w, AzureResources(w, include_subscriptions="002"), location
-    )
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    storage_permission_mappings = azure_resource_permission._map_storage(AzureResource(resource_id))
-    assert len(storage_permission_mappings) == 0
-
-
-def test_get_role_assignments_with_spn_and_blob_permission(mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    location = ExternalLocations(w, MockBackend(), "ucx")
-    resource_id = "subscriptions/002/resourceGroups/rg1/storageAccounts/sto2"
-    installation = MockInstallation()
-    azure_resource_permission = AzureResourcePermissions(
-        installation, w, AzureResources(w, include_subscriptions="002"), location
-    )
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    storage_permission_mappings = azure_resource_permission._map_storage(AzureResource(resource_id))
-    assert len(storage_permission_mappings) == 2
-    for storage_permission_mapping in storage_permission_mappings:
-        assert storage_permission_mapping.prefix == "abfss://container3@sto2.dfs.core.windows.net/"
-        assert storage_permission_mapping.principal == "disNameuser3"
-        assert storage_permission_mapping.privilege == "WRITE_FILES"
-        assert storage_permission_mapping.client_id == "appIduser3"
-
-
-def test_save_spn_permissions_no_valid_storage_accounts(caplog, mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    rows = {"SELECT \\* FROM ucx.external_locations": [["abfss://continer1@sto3.dfs.core.windows.net/folder1", 1]]}
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    backend = MockBackend(rows=rows)
-    location = ExternalLocations(w, backend, "ucx")
-    installation = MockInstallation()
-    azure_resource_permission = AzureResourcePermissions(
-        installation, w, AzureResources(w, include_subscriptions="002"), location
-    )
-    azure_resource_permission.save_spn_permissions()
-    assert [rec.message for rec in caplog.records if "No storage account found in current tenant" in rec.message]
-
-
-def test_save_spn_permissions_valid_storage_accounts(caplog, mocker, az_token):
-    w = create_autospec(WorkspaceClient)
-    rows = {"SELECT \\* FROM ucx.external_locations": [["abfss://continer1@sto2.dfs.core.windows.net/folder1", 1]]}
-    mocker.patch("databricks.sdk.core.ApiClient.do", side_effect=get_az_api_mapping)
-    backend = MockBackend(rows=rows)
-    location = ExternalLocations(w, backend, "ucx")
-    installation = MockInstallation()
-    azure_resource_permission = AzureResourcePermissions(
-        installation, w, AzureResources(w, include_subscriptions="002"), location
-    )
-    azure_resource_permission.save_spn_permissions()
-    installation.assert_file_written(
-        'azure_storage_account_info.csv',
-        [
-            {
-                'client_id': 'appIduser3',
-                'prefix': 'abfss://container3@sto2.dfs.core.windows.net/',
-                'principal': 'disNameuser3',
-                'privilege': 'WRITE_FILES',
-            },
-            {
-                'client_id': 'appIduser3',
-                'prefix': 'abfss://container3@sto2.dfs.core.windows.net/',
-                'principal': 'disNameuser3',
-                'privilege': 'WRITE_FILES',
-            },
-        ],
-    )
+from . import workspace_client_mock
 
 
 def test_azure_spn_info_without_secret():
