+++ tests/unit/azure/azure/mappings.json
